# Improved performance of creating BulgeGraph