@Bernhard10 2023-03-13
@domen13 Now that I think about it, I'm wondering if a linear runtime could be achieved by iterating over the stems (defines starting with 'y' in this case) and checking the attached bulges instead of iterating over the bulges and checking if they are  attached to the same stems (edges).